ref(participants-pane) Use new button component (#11913)
