+++ react/features/base/components/index.js
@@ -1,4 +1,3 @@
 export { default as ContextMenu } from './context-menu/ContextMenu';
 export { default as ContextMenuItemGroup } from './context-menu/ContextMenuItemGroup';
 export { default as ListItem } from './participants-pane-list/ListItem';
-export { default as QuickActionButton } from './buttons/QuickActionButton';
